Update make-your-own-image.md
I thought I had to install an image, so I wasted a ton of time on this, just to learn that the reason I didn't see the WiTi Board on my Wi-Fi network once powered is that Wi-Fi is off by default. Hopefully my edits will help with clarity this document and save other some time.